New Module: Deep Learning support with Keras
JIRA:

This is an introductory commit to supporting Deep Learning in MADlib.
The goal of this feature is to make it easy for data scientists to
develop models using deep learning on data residing in
Postgres/Greenplum database. We do not want users to have to move their
data from their database to another place in order to do deep learning.

The current implementation stands with the following requirements:
- The database host has Tensorflow and Keras set up.
- Users must use MADlib's minibatch_preprocessor_dl() module to
  pre-process their data, since this module expects the data to be in
  the format output by minibatch_preprocessor_dl().
- The model architecture is expected to be stored in a different table,
  which is passed along as a param to this module.
- If there are no GPUs in the host, the use_gpu param must be set to
  false while calling this module.

There are a number of hard assumptions that we make, which must be
addressed in other JIRAs (not a complete list).
TODO JIRAs:
- Postgres support for Deep learning ().
- There is a hard assumption regarding the number of GPUs per host,
  which should be fixed as part of another JIRA ().
- Documentation for the module ().
- Generalize and refactor predict.

Co-authored-by: Domino Valdano <dvaldano@pivotal.io>
Co-authored-by: Nikhil Kak <nkak@pivotal.io>
Co-authored-by: Rahul Iyer <riyer@apache.org>
Co-authored-by: Arvind Sridhar <asridhar@pivotal.io>
Co-authored-by: Omer Arap <oarap@pivotal.io>