DL: Add training for multiple models
This commit adds a new function to train multiple models in parallel
with model hopping method supported for Greenplum DB only.

Model hopping method involves the following steps:

- Train models in parallel for a single epoch using the data local
to each segment.
- Move the models to the next segment in round-robin fashion.

This method ensures that all of the models visit the entire dataset,
which eliminates the need to average the model at the end.

This commit also fixes an issue with the regular fit function where
excessive amounts of threads were being created and left over by keras
sessions. It is fixed by reusing the same session and the same
computational graph throughout the process.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
Co-authored-by: Nandish Jayaram <njayaram@apache.org>
Co-authored-by: Nikhil Kak <nkak@pivotal.io>
Co-authored-by: Orhan Kislal <okislal@apache.org>