DL: Check for NULL in class_values for fit_multiple_model
For, already one-hot encoded data, the class_values is set to NULL by
the input preprocessor. Prior to this commit, the function
`madlib_keras_fit_multiple_model()`, assumed the class_values will
always be present and failed incase of class_values being passed in as
NULL. This commit fixes this issue and adds a dev-check test for the same.

Authored-by: Ekta Khanna <ekhanna@pivotal.io>