DL: Improve performance of mini-batch preprocessor
This commit adds the following optimizations to minibatch preprocessor:

- Skip normalization if normalization contant is 1.0

- Split the batching query to generate buffer_id's (based of row_id)
without moving around any data. Previously, calling `ROW_NUMBER()
OVER()` to add row_id's to the table was causing the data to be gathered
on the master node and then numbering the rows, which for large datasets
would be taking most of the time.

- Separate out the JOIN (called for even distribution) as well as
converting to bytea from the batching query. This avoids, any VMEM limit
issues.

- num_buffers gets rounded up to the nearest multiple of num_segments
for even distribution across buffers on segments.

- Add new C function `array_to_bytea()` to convert array to bytea. This
speeds up the query significantly.

Additionally, this commit adds a new function `plpy_execute()` that
prints EXPLAIN plans and execution time for debugging.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>