DL: Avoid constant folding of weights in GPDB6 plan
JIRA: MADLIB-1405

For versions >=GPDB6, previously, for queries called with the initial
weights value passed in, the query plan for it would have create custom
plans with weights embedded in the plan itself.  This meant that the
query plan size would also include the size of these weights, bloating
it up to hit the 1GB limit when dispatching the query plan to segments,
leading to OOM for large weights.

In GPDB, for PREPARE plans, there is a threshold of 5 attempts to create
custom plans(constant folding the passed in params) for execution and
then it uses a generic plan(not constant folding the passed in params)
for all the subsequent executions.  Therefore, to avoid GPDB6 from
creating custom plans when passing in weights, the queries(with weights)
is executed with DUMMY weights for 5 time, prior to calling it with the
actual weights.