DL: Remove final function for fit multiple
We don't really need the final function since we don't care about the
image count.
Based on our experiments, we noticed that at the end of each hop, fit
multiple would start using a lot of memory (up from 80 GB to ~ 140 GB on
our gcp cluster) and we attributed that memory spike to the final function.