DL: Use unlogged tables for fit multiple
GPDB6 supports UNLOGGED tables. From gpdb documentation
https://gpdb.docs.pivotal.io/6-0/ref_guide/sql_commands/CREATE_TABLE_AS.html

```
Data written to unlogged tables is not written to the write-ahead (WAL)
log, which makes them considerably faster than ordinary tables
```

We saw speed improvements because of using unlogged tables on our gcp
cluster with GPDB6. But, we do not want the final model output table,
summary table and the info table to be unlogged. This commit ensures for
all intermediate iterations, except the very final hop in the last
iteration, the model output table is created as `UNLOGGED` for
performance. Only for the last hop in the last iteration, it is created
as a persistent table.

This commit also adds a test to validate model output, summary and info
tables are persistent(not unlogged) and adds a new utility function to
check if the gpdb version is 6 and up.

Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>