DL: Create model table as unlogged for intermediate iteration
We see a speedup for each hop in an iteration when creating tables as
`UNLOGGED`. But, we do not want the final model output table, summary
table and the info table to be unlogged. This commit ensures for all
intermediate iterations, except the very final hop in the last
iteration, the model output table is created as `UNLOGGED` for
performance. Only for the last hop in the last iteration, it is created
as a persistent table.

Also, added a test to validate model output, summary and info tables are
persistent(not unlogged).

Co-authored-by: Nikhil Kak <nkak@pivotal.io>