Add MinWarning to remove extraneous INFO messages
JIRA: MADLIB-1453

We might also want to add it to PythonFunction macro, and
automatically wrap every function call with:
 with MinWarning('warning')
This is what we do for AOControl, which makes sense to me.  But
before enabling that we would need to check to make sure nothing
in MADlib is counting on INFO statements showing up by default.
For now, I'm just importing it by default so that it can easily
be used along with AOControl in the .sql_in files.  This is much
better than putting decorators on all of our functions in the
.py_in files, as the latter will keep turning this GUC on and
off every time a function is entered or exited.