DL: Use GD instead of SD for storing session/model
JIRA: MADLIB-1438

For fit, since we can reuse the same session for all calls to fit and
evaluate we use GD instead of SD. The session gets reset at the final
call to evaluate.
Only the very final evaluate should clear the
session which may not necessarily be the training evaluate, so changed the
code to take care of this scenario

For fit_multiple, we need to clear the session for each hop(last row of
each call) when calling fit/evaluate step functions.

Like SD, this code also assumes that fit and evaluate run in the same
process so GD can be shared across function calls.

Co-authored-by: Ekta Khanna <ekhanna@vmware.com>