DL: Update error msg for missing input_shape
JIRA: MADLIB-1438

Previously, `keras` would error out if the first layer didn't have an
input shape but with `tf.keras` there is way to create a layer without
an input shape.(https://keras.io/examples/vision/mnist_convnet/). In
this case, the model JSON does not have it either. We need the input
shape for validation and also we cannot set model weights without the
input shape being added to the first layer.

Co-authored-by: Nikhil Kak <nkak@vmware.com>