Replace FULL JOIN with LEFT JOIN
I think I made this change while merging in the caching changes,
to get some tests with caching enabled to pass.  But it seems
to have added an extra Redistribute Motion which can cause an
intermittant crash.  Retrospectively, I don't think it was
actually necessary as there shouldn't be any dist keys in the
source table that don't match keys in the input table; reverting
it to LEFT JOIN so we can test it again.

Also:

- Remove extra call to query_weights()
- Disable unused function definition for gpdb5 in fit mult testfile