fix caching issue & fix gpdb5 tests
Reverting from FULL JOIN to LEFT JOIN works fine for most cases, but there is one edge case where it breaks:

If caching is enabled & num msts < num segs, then there
will be some segments without models that need to
have their cache populated on the very first hop, and cleared on the very last.

These won't even be the same set of segments, since
the models will have hopped to different segments many
times between the first and last call to run_training.

To fix this, we can insert an empty row in the model_input table just for these 2 special cases.
The FULL JOIN was an easier solution, but this should
work just as well, and can allow us to get by with just
a LEFT JOIN.

Also:  PRIMARY KEY constraint on input table has been
reduced from mst_key to just dist_key, so that we can
have NULL mst_key columns in the empty rows.  I think
this was needed for the GROUP BY in the UDA, but
it should no longer be needed for the UDF.